Removed console spam

+++ GameEngine/collider.cpp
@@ -17,18 +17,16 @@ collider::~collider()
 
 void collider::updateColliders()
 {
-    int posx = m_p_parent->get_position().x + m_position_offset->x;
-    int posy = m_p_parent->get_position().y + m_position_offset->y;
+    float posx = m_p_parent->get_position().x + m_position_offset->x;
+    float posy = m_p_parent->get_position().y + m_position_offset->y;
 
     for(int i = 0; i < m_otherColliders.size(); i++ )
     {
-        int diffx = posx - m_otherColliders[i]->getParent()->get_position().x;
-        int diffy = posy - m_otherColliders[i]->getParent()->get_position().y;
+        float diffx = posx - m_otherColliders[i]->getParent()->get_position().x;
+        float diffy = posy - m_otherColliders[i]->getParent()->get_position().y;
 
-        int mag = (diffx * diffx) + (diffy * diffy);
+        float mag = (diffx * diffx) + (diffy * diffy);
         mag = sqrt(mag);
-        //printf(std::to_string(mag).c_str());
-        //printf("\n");
 
         if(mag < m_radius + m_otherColliders[i]->getRadius())
         {
